Move CommandMigrateConfigError and ServerNameModifiedError to Synapse
